restructured spec files
